unary minus on an unsigned type is still a positive (unsigned) number... offsetting to 2^64-padding was almost certainly not the intention.
svn:revision:66429
svn:branch:trunk
svn:account:brlcad

+++ src/librt/primitives/bot/gct_decimation/auxiliary/mm.c
@@ -361,8 +361,8 @@ static mmBlock *mmBlockResolveChunk(void *p, mmBlock *root)
 static void mmAlignRelayFree(void (*relayfree)(void *head, void *v, size_t bytes), void *relayvalue, void *v, size_t bytes)
 {
     mmAlign *malign;
-    malign = (mmAlign *)ADDRESS(v, -sizeof(mmAlign));
-    relayfree(relayvalue, ADDRESS(v, -malign->padding), bytes);
+    malign = (mmAlign *)ADDRESS(v, -(intptr_t)sizeof(mmAlign));
+    relayfree(relayvalue, ADDRESS(v, -(intptr_t)malign->padding), bytes);
 }
